Add files via upload
Once the application is initialized, the jinja template renders the webpage and looks for the following statements, {% if session %}, {% else %} and {% endif %}. Then redirects the uri to logged_out.html. On this landing page, a user can navigate the application or test login via Okta OAuth 2.0 API Integration. Using Okta OAuth 2.0 integration is a seamless experience. Checkout the README for more information.